fix: online notifications do not push to herself.

+++ removeall.sh
@@ -15,11 +15,14 @@
 # limitations under the License.
 
 make stop
+if [ $? -eq 0 ]; then
+    docker-compose -f docker-compose-a.yml down
+    wait $!
 
-docker-compose -f docker-compose-a.yml down
-wait $!
-
-rm -rf components/kafka/ components/redis/ components/mongodb/
-rm logs/*
-docker-compose -f docker-compose-a.yml up -d
-make start
+    if [ $? -eq 0 ]; then
+        rm -rf components/kafka/ components/redis/ components/mongodb/
+        rm logs/*
+        docker-compose -f docker-compose-a.yml up -d
+        make start
+    fi
+fi
